# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tahr Desk",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/login/access-token": {
      "post": {
        "tags": ["login"],
        "summary": "Login Access Token",
        "operationId": "login-login_access_token",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login-login_access_token"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/login/test-token": {
      "post": {
        "tags": ["login"],
        "summary": "Test Token",
        "operationId": "login-test_token",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/public/services/{slug}": {
      "get": {
        "tags": ["public"],
        "summary": "Get Service",
        "operationId": "public-get_service",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceCategoryPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/cases": {
      "post": {
        "tags": ["public"],
        "summary": "Create Case",
        "operationId": "public-create_case",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_public-create_case"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/cases/lookup": {
      "post": {
        "tags": ["public"],
        "summary": "Lookup Case",
        "operationId": "public-lookup_case",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseCredentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CasePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/cases/messages": {
      "post": {
        "tags": ["public"],
        "summary": "Create Public Message",
        "operationId": "public-create_public_message",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_public-create_public_message"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseMessagePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/attachments/download": {
      "post": {
        "tags": ["public"],
        "summary": "Download Attachment",
        "operationId": "public-download_attachment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachmentDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttachmentDownloadRequest": {
        "properties": {
          "reference": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Reference"
          },
          "reporter_email": {
            "type": "string",
            "format": "email",
            "title": "Reporter Email"
          },
          "attachment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Attachment Id"
          }
        },
        "type": "object",
        "required": ["reference", "reporter_email", "attachment_id"],
        "title": "AttachmentDownloadRequest"
      },
      "AttachmentPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type"
          },
          "byte_count": {
            "type": "integer",
            "title": "Byte Count"
          },
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_name",
          "media_type",
          "byte_count",
          "width",
          "height"
        ],
        "title": "AttachmentPublic"
      },
      "Body_login-login_access_token": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": ["username", "password"],
        "title": "Body_login-login_access_token"
      },
      "Body_public-create_case": {
        "properties": {
          "submission_id": {
            "type": "string",
            "format": "uuid",
            "title": "Submission Id"
          },
          "reporter_name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Reporter Name"
          },
          "reporter_email": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Reporter Email"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "subject": {
            "type": "string",
            "maxLength": 180,
            "minLength": 5,
            "title": "Subject"
          },
          "description": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 20,
            "title": "Description"
          },
          "location_text": {
            "type": "string",
            "maxLength": 240,
            "minLength": 3,
            "title": "Location Text"
          },
          "photos": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "contentMediaType": "application/octet-stream"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photos"
          }
        },
        "type": "object",
        "required": [
          "submission_id",
          "reporter_name",
          "reporter_email",
          "category_id",
          "subject",
          "description",
          "location_text"
        ],
        "title": "Body_public-create_case"
      },
      "Body_public-create_public_message": {
        "properties": {
          "reference": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Reference"
          },
          "reporter_email": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Reporter Email"
          },
          "body_markdown": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Body Markdown"
          },
          "photos": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "contentMediaType": "application/octet-stream"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photos"
          }
        },
        "type": "object",
        "required": ["reference", "reporter_email", "body_markdown"],
        "title": "Body_public-create_public_message"
      },
...
      "CaseCredentials": {
        "properties": {
          "reference": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Reference"
          },
          "reporter_email": {
            "type": "string",
            "format": "email",
            "title": "Reporter Email"
          }
        },
        "type": "object",
        "required": ["reference", "reporter_email"],
        "title": "CaseCredentials"
      },
      "CaseEventPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "from_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "to_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_type",
          "summary",
          "from_status",
          "to_status",
          "created_at"
        ],
        "title": "CaseEventPublic"
      },
      "CaseMessagePublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "author_label": {
            "type": "string",
            "title": "Author Label"
          },
          "body_html": {
            "type": "string",
            "title": "Body Html"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentPublic"
            },
            "type": "array",
            "title": "Attachments"
          }
        },
        "type": "object",
        "required": ["id", "author_label", "body_html", "created_at"],
        "title": "CaseMessagePublic"
      },
...
      "CasePublic": {
        "properties": {
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "category": {
            "$ref": "#/components/schemas/ServiceCategoryPublic"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "location_text": {
            "type": "string",
            "title": "Location Text"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentPublic"
            },
            "type": "array",
            "title": "Attachments"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/CaseMessagePublic"
            },
            "type": "array",
            "title": "Messages"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/CaseEventPublic"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "reference",
          "category",
          "subject",
          "description",
          "location_text",
          "status",
          "created_at",
          "updated_at",
          "attachments",
          "messages",
          "events"
        ],
        "title": "CasePublic"
      },
      "CaseReceipt": {
        "properties": {
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "reporter_email_masked": {
            "type": "string",
            "title": "Reporter Email Masked"
          }
        },
        "type": "object",
        "required": [
          "reference",
          "status",
          "created_at",
          "reporter_email_masked"
        ],
        "title": "CaseReceipt"
      },
      "CaseStatus": {
        "type": "string",
        "enum": [
          "submitted",
          "triaged",
          "in_progress",
          "waiting_on_reporter",
          "resolved",
          "closed"
        ],
        "title": "CaseStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ServiceCategoryPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "guidance_html": {
            "type": "string",
            "title": "Guidance Html"
          },
          "response_target_hours": {
            "type": "integer",
            "title": "Response Target Hours"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "summary",
          "guidance_html",
          "response_target_hours",
          "sort_order",
          "is_active"
        ],
        "title": "ServiceCategoryPublic"
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": ["access_token"],
        "title": "Token"
      },
...
      "UserPublic": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "is_superuser": {
            "type": "boolean",
            "title": "Is Superuser"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "email",
          "is_active",
          "is_superuser",
          "full_name",
          "id",
          "role",
          "created_at"
        ],
        "title": "UserPublic"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}